refactor: :recycle: OAuth 인수테스트에 restdocs가관련 코드 추가
