+++ be/src/docs/asciidoc/index.adoc
@@ -15,6 +15,12 @@ notification-api-docs
 
 include::{docdir}/review/review.adoc[]
 
+
 == 음식점 API
 
 include::{docdir}/store/store.adoc[]
+
+== 로그인 API
+
+include::{docdir}/auth/auth.adoc[]
+
