+++ be/src/test/java/com/jjikmuk/sikdorak/acceptance/auth/OAuthAcceptanceTest.java
@@ -1,5 +1,6 @@
-package com.jjikmuk.sikdorak.acceptance;
+package com.jjikmuk.sikdorak.acceptance.auth;
 
+import com.jjikmuk.sikdorak.acceptance.InitAcceptanceTest;
 import com.jjikmuk.sikdorak.common.mock.OAuthMocks;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -11,16 +12,18 @@ import org.springframework.test.context.TestPropertySource;
 
 import java.io.IOException;
 
+import static com.jjikmuk.sikdorak.acceptance.auth.OAuthSnippet.*;
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.notNullValue;
-import static org.hamcrest.Matchers.startsWith;
+import static org.hamcrest.Matchers.*;
+import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
 @AutoConfigureWireMock(port = 0)
 @TestPropertySource(properties = {
         "oauth.kakao.service.token_url=http://localhost:${wiremock.server.port}",
         "oauth.kakao.service.api_url=http://localhost:${wiremock.server.port}"
 })
-public class LoginAcceptanceTest extends InitAcceptanceTest {
+@DisplayName("OAuth 인수테스트")
+public class OAuthAcceptanceTest extends InitAcceptanceTest {
 
     @BeforeEach
     void setWireMockResponse() throws IOException {
@@ -31,7 +34,8 @@ public class LoginAcceptanceTest extends InitAcceptanceTest {
     @Test
     @DisplayName("유저가 로그인 버튼을 눌렀을 경우 카카오 로그인 페이지 url을 반환해야 한다.")
     void redirect_to_social_login_page() {
-        given()
+        given(this.spec)
+                .filter(document(DEFAULT_RESTDOC_PATH))
                 .when()
                 .redirects().follow(false)
                 .get("/api/oauth/login")
@@ -42,8 +46,11 @@ public class LoginAcceptanceTest extends InitAcceptanceTest {
 
     @Test
     @DisplayName("유저로부터 로그인 요청이 들어오면 유저의 정보를 저장하고 토큰을 발급한다.")
-    void new_user_login_success() {
-        given()
+    void user_login_success() {
+        given(this.spec)
+                .filter(document(DEFAULT_RESTDOC_PATH,
+                        LOGIN_SUCCESS_REQUEST_SNIPPET,
+                        LOGIN_SUCCESS_RESPONSE_SNIPPET))
                 .queryParam("code", "code")
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when()
