+++ be/src/test/java/com/jjikmuk/sikdorak/common/mock/OAuthMocks.java
@@ -6,6 +6,7 @@ import org.springframework.http.MediaType;
 
 import java.io.IOException;
 
+import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
 import static java.nio.charset.Charset.defaultCharset;
 import static org.springframework.util.StreamUtils.copyToString;
@@ -24,6 +25,7 @@ public class OAuthMocks {
 
     public static void setupMockUserInfoResponse() throws IOException{
         stubFor(WireMock.get(WireMock.urlEqualTo("/v2/user/me"))
+                .withHeader("Authorization",equalTo("bearer accessToken"))
                 .willReturn(WireMock.aResponse()
                         .withStatus(HttpStatus.OK.value())
                         .withHeader("Content-Type", MediaType.APPLICATION_JSON_VALUE)
